+++ lib/src/app/supplemental/asymmetric_view.dart
@@ -14,8 +14,10 @@
 
 import 'package:flutter/material.dart';
 
-import 'package:shrine_mvc/src/home/model/product.dart';
-import 'package:shrine_mvc/src/app/supplemental/product_columns.dart';
+import 'package:shrine_mvc/src/model.dart' show Product;
+
+import 'package:shrine_mvc/src/app/supplemental/product_columns.dart'
+    show OneProductCardColumn, TwoProductCardColumn;
 
 class AsymmetricView extends StatelessWidget {
   const AsymmetricView({Key key, this.products}) : super(key: key);
@@ -35,7 +37,8 @@ class AsymmetricView extends StatelessWidget {
     // some kinda awkward math so we use _evenCasesIndex and _oddCasesIndex as
     // helpers for creating the index of the product list that will correspond
     // to the index of the list of columns.
-    return List<Container>.generate(_listItemCount(products.length), (int index) {
+    return List<Container>.generate(_listItemCount(products.length),
+        (int index) {
       double width = .59 * MediaQuery.of(context).size.width;
       Widget column;
       if (index % 2 == 0) {
@@ -43,9 +46,7 @@ class AsymmetricView extends StatelessWidget {
         final int bottom = _evenCasesIndex(index);
         column = TwoProductCardColumn(
           bottom: products[bottom],
-          top: products.length - 1 >= bottom + 1
-            ? products[bottom + 1]
-            : null,
+          top: products.length - 1 >= bottom + 1 ? products[bottom + 1] : null,
         );
         width += 32.0;
       } else {
@@ -79,8 +80,8 @@ class AsymmetricView extends StatelessWidget {
 
   int _listItemCount(int totalItems) {
     return (totalItems % 3 == 0)
-      ? totalItems ~/ 3 * 2
-      : (totalItems / 3).ceil() * 2 - 1;
+        ? totalItems ~/ 3 * 2
+        : (totalItems / 3).ceil() * 2 - 1;
   }
 
   @override