+++ lib/src/app/supplemental/product_columns.dart
@@ -14,8 +14,10 @@
 
 import 'package:flutter/material.dart';
 
-import 'package:shrine_mvc/src/home/model/product.dart';
-import 'package:shrine_mvc/src/app/supplemental/product_card.dart';
+import 'package:shrine_mvc/src/model.dart' show Product;
+
+import 'package:shrine_mvc/src/app/supplemental/product_card.dart'
+    show ProductCard;
 
 class TwoProductCardColumn extends StatelessWidget {
   const TwoProductCardColumn({
@@ -27,16 +29,19 @@ class TwoProductCardColumn extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return LayoutBuilder(builder: (BuildContext context, BoxConstraints constraints) {
+    return LayoutBuilder(
+        builder: (BuildContext context, BoxConstraints constraints) {
       const double spacerHeight = 44.0;
 
-      final double heightOfCards = (constraints.biggest.height - spacerHeight) / 2.0;
-      final double availableHeightForImages = heightOfCards - ProductCard.kTextBoxHeight;
+      final double heightOfCards =
+          (constraints.biggest.height - spacerHeight) / 2.0;
+      final double availableHeightForImages =
+          heightOfCards - ProductCard.kTextBoxHeight;
       // Ensure the cards take up the available space as long as the screen is
       // sufficiently tall, otherwise fallback on a constant aspect ratio.
       final double imageAspectRatio = availableHeightForImages >= 0.0
-        ? constraints.biggest.width / availableHeightForImages
-        : 49.0 / 33.0;
+          ? constraints.biggest.width / availableHeightForImages
+          : 49.0 / 33.0;
 
       return ListView(
         physics: const ClampingScrollPhysics(),
@@ -44,13 +49,13 @@ class TwoProductCardColumn extends StatelessWidget {
           Padding(
             padding: const EdgeInsetsDirectional.only(start: 28.0),
             child: top != null
-              ? ProductCard(
-                  imageAspectRatio: imageAspectRatio,
-                  product: top,
-                )
-              : SizedBox(
-                  height: heightOfCards > 0 ? heightOfCards : spacerHeight,
-                ),
+                ? ProductCard(
+                    imageAspectRatio: imageAspectRatio,
+                    product: top,
+                  )
+                : SizedBox(
+                    height: heightOfCards > 0 ? heightOfCards : spacerHeight,
+                  ),
           ),
           const SizedBox(height: spacerHeight),
           Padding(