+++ lib/src/view.dart
@@ -23,8 +23,6 @@
 
 export 'package:mvc_application/view.dart';
 
-export 'package:shrine_mvc/src/home/view/home.dart';
-
 export 'package:shrine_mvc/src/app/view/app.dart';
 
 export 'package:shrine_mvc/src/app/view/login.dart';
@@ -33,5 +31,13 @@ export 'package:shrine_mvc/src/app/view/colors.dart';
 
 export 'package:shrine_mvc/src/app/supplemental/cut_corners_border.dart';
 
+export 'package:shrine_mvc/src/home/view/home.dart';
 
+export 'package:shrine_mvc/src/home/view/backdrop.dart';
+
+export 'package:shrine_mvc/src/home/view/expanding_bottom_sheet.dart';
+
+export 'package:shrine_mvc/src/home/view/category_menu_page.dart';
+
+export 'package:shrine_mvc/src/app/supplemental/asymmetric_view.dart';