+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.3.1
+ August 30, 2019
+- i18n Internationalization capability
+
+## 0.3.0 
+- SetState(builder: (context, _) =>
+
 ## 0.2.0
  August 22, 2019
 - AppStateModel() in the View, ShrineApp